Modifications to enable handling of slave DOFs and related DOF managers (rigid arm nodes, hanging nodes) in Load Balancing Engine.
The modification ensures the availability of slave and corresponding 
master DOFs (in terms of DOF managers) on the same partition. As slave 
DOFs are expressed in terms of (linear) combination of master DOFs, the 
master DOFs (and corresponding of managers) should be available on any 
partition with slave DOFs to accommodate slave contributions.

Changes:
- Modification at load balancer itself
- Added context store & restore services for individual DOF representations
- Slave DOF class re-casted to use index references to masters 
  instead of pointers to individual DOFs to enable remote transfer
- updateLocalNumbering methods extended to individual DOFs
  and implemented using functors.