Added: beReducedVectorFormOfStrain to convert from matrix form to reduced Voight form
Renamed: beReducedVectorForm to beReducedVectorFormOfStress to make it clear that it expects the argument to be stress
Renamed: beMatrixForm to beMatrixFormOfStress  to make it clear that it expects the argument to be stress

Maybe one should include some new stress-strain classes (which includes the above methods) to reduce the number of methods in FloatArray and FloatMatrix.
One can also add methods for converting between different stress and strain measures which becomes important in large deformation theory.