Major revision of vtkXMLExportModule. The export is now clearly divided into two steps: 1) export/extract all the data and store it in a VTKPiece 2) Write the VTKPiece to file.
By doing it this way, output of composite elements can be handled with the same code. Only that each element is responsible of setting up an VTKPiece.
Hopefully, the code is easier to read and maintain now.

TODO:
The revision has not been tested with the VTK module so there are most likely some errors.
There are some things that should be improved/revised see todo's in the code.
Members in the VTKPiece are public, they should be made private and methods for accessing them should be implemented.