Fixed size math for LargeStrainMaterMaterial
Note: The old expressions didn't add up: matrice sizes for the first PK was a mismatch of voigt and matrix form.
This means that there was no way the old code was working correctly.
With the fixed size matrices, this became a compile error, and I tried to remedy it in a way that made sense.