Move coordinate from Node to DofManager.
Almost all code in OOFEM assumes that dofmanagers always have coordinates anyway.
Moving this to the DofManager base class eliminates the overhead of a virtual call, which is very important as this is called a lot.
ElementDofManagers could even, if necessary, just store the element midpoint as a meaningful coordinate.